+++ tests/features/nav_nested/nav_nested.c
@@ -18,7 +18,7 @@ int main(int argc, char** argv) {
     char match_;
     SubRecord_ = (struct SubRecord_){.t0 = (struct String){.ptr = "Hello", .len = 5}};
     text_ = SubRecord_;
-    milone_assert((text_.t0.len == 5), (struct String){.ptr = "nav_nested/nav_nested.milone", .len = 28}, 8, 2);
+    milone_assert((((int32_t)text_.t0.len) == 5), (struct String){.ptr = "nav_nested/nav_nested.milone", .len = 28}, 8, 2);
     value_ = 42;
     milone_assert((value_ == 42), (struct String){.ptr = "nav_nested/nav_nested.milone", .len = 28}, 11, 35);
     match_ = 0;
